Add files via upload
To someone who's never coded before, the concept of creating a website from scratch -- layout, design using a responsive web design. The goal of this small project is to make a responsive webpage by implementing basics of HTML, CSS and  JavaScript with hamburger menu and sticky navbar using bootstrap and Visual Studio Code as code editor.